remove: duplicate tests/ field.rs, pairing.rs
