+++ curves/src/tests/mod.rs
@@ -1,5 +1,3 @@
 pub mod curve;
 pub mod engine;
-pub mod field;
 pub mod field_halo2curves;
-pub mod pairing;
